# openapi/swagger.yaml
components:
  schemas:
...
    internal_httpapi.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    internal_httpapi.OAuthErrorResponse:
      properties:
        error:
          type: string
        error_description:
          type: string
        pending_token:
          type: string
      type: object
    internal_httpapi.OAuthTokenResponse:
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
        refresh_token:
          type: string
        token_type:
          type: string
      type: object
...
    sudoStream_internal_access.LibraryType:
      enum:
      - film
      - series
      - music
      - photos
      - other
      type: string
      x-enum-varnames:
      - LibraryTypeFilm
      - LibraryTypeSeries
      - LibraryTypeMusic
      - LibraryTypePhotos
      - LibraryTypeOther
...
    sudoStream_internal_metadata.FilenameHint:
      properties:
        episode:
          type: integer
        season:
          type: integer
      type: object
    sudoStream_internal_metadata.MetadataResponse:
      properties:
        display_name:
          type: string
        effective:
          $ref: '#/components/schemas/sudoStream_internal_metadata.VideoFields'
        filename_hint:
          $ref: '#/components/schemas/sudoStream_internal_metadata.FilenameHint'
        library_type:
          $ref: '#/components/schemas/sudoStream_internal_access.LibraryType'
        overridden_by:
          type: string
        path:
          type: string
        probed_at:
          type: string
        source:
          $ref: '#/components/schemas/sudoStream_internal_metadata.SourceFields'
        uses_metadata_for_display:
          type: boolean
      type: object
    sudoStream_internal_metadata.PatchRequest:
      properties:
        fields:
          additionalProperties:
            $ref: '#/components/schemas/sudoStream_internal_metadata.jsonValue'
          type: object
        target:
          $ref: '#/components/schemas/sudoStream_internal_metadata.PatchTarget'
      type: object
    sudoStream_internal_metadata.PatchTarget:
      enum:
      - override
      - file
      type: string
      x-enum-varnames:
      - PatchTargetOverride
      - PatchTargetFile
...
    sudoStream_internal_metadata.SourceFields:
      properties:
        original:
          $ref: '#/components/schemas/sudoStream_internal_metadata.VideoFields'
        override:
          $ref: '#/components/schemas/sudoStream_internal_metadata.VideoFields'
      type: object
    sudoStream_internal_metadata.VideoFields:
      properties:
        actors:
          items:
            type: string
          type: array
          uniqueItems: false
        audio_channels:
          type: integer
        audio_codec:
          type: string
        audio_languages:
          items:
            type: string
          type: array
          uniqueItems: false
        bitrate:
          type: integer
        composer:
          type: string
        content_rating:
          type: string
        country:
          type: string
        description:
          type: string
        directors:
          items:
            type: string
          type: array
          uniqueItems: false
        duration_seconds:
          type: number
        ed2k_hash:
          type: string
        episode:
          type: integer
        episode_title:
          type: string
        frame_rate:
          type: string
        genres:
          items:
            type: string
          type: array
          uniqueItems: false
        hash_file_size:
          description: size when hashes were computed
          type: integer
        height:
          type: integer
        imdb_id:
          type: string
        language:
          type: string
        movie_hash:
          description: System-owned identity hashes (scan-time; never overwritten;
            not PATCH-editable).
          type: string
        original_title:
          type: string
        producers:
          items:
            type: string
          type: array
          uniqueItems: false
        release_date:
          type: string
        season:
          type: integer
        show:
          type: string
        sort_title:
          type: string
        studio:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
        tmdb_id:
          type: string
        tvdb_id:
          type: string
        tvmaze_id:
          type: string
        video_codec:
          type: string
        width:
          type: integer
        writers:
          items:
            type: string
          type: array
          uniqueItems: false
        year:
          type: integer
      type: object
    sudoStream_internal_metadata.jsonValue:
      type: object
...
info:
  description: HTTP API for browsing and streaming media files.
  title: sudoStream API
  version: 0.35.0
openapi: 3.1.0
paths:
...
  /api/metadata/{path}:
    get:
      description: Returns cached original and override metadata for a video path.
      parameters:
      - description: Path under media root
        in: path
        name: path
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_metadata.MetadataResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Get video metadata
      tags:
      - media
    patch:
      description: Updates per-file override metadata or writes embedded file tags
        via ffmpeg.
      parameters:
      - description: Path under media root
        in: path
        name: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/sudoStream_internal_metadata.PatchRequest'
                description: Metadata patch
                summary: body
        description: Metadata patch
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sudoStream_internal_metadata.MetadataResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.ErrorResponse'
          description: Internal Server Error
      summary: Patch video metadata
      tags:
      - media
  /api/oauth/token:
    post:
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.OAuthTokenResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.OAuthErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_httpapi.OAuthErrorResponse'
          description: Unauthorized
      summary: OAuth2 token
      tags:
      - oauth
...
servers:
- description: Local development server
  url: http://localhost:8080